Backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiolimiter==1.2.1
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.43.1
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
assemblyai==0.26.0
astrapy==1.5.2
asttokens==3.0.0
async-lru==2.0.4
asyncer==0.0.5
attrs==24.3.0
Authlib==1.3.1
babel==2.16.0
backoff==2.2.1
bce-python-sdk==0.9.25
bcrypt==4.0.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
boto3==1.34.162
botocore==1.34.162
branca==0.8.1
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.0
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.0.1
chroma-hnswlib==0.7.6
clevercsv==0.8.3
click==8.1.8
cohere==5.13.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.0.7
cryptography==42.0.8
cycler==0.11.0
dataclasses-json==0.6.7
datasets==2.2.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.9
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
docker==7.1.0
docstring-parser==0.15
dspy==2.5.29
duckdb==1.1.3
ecdsa==0.19.0
elastic-transport==8.17.0
elasticsearch==8.17.0
email_validator==2.2.0
emoji==2.14.1
eval_type_backport==0.2.2
eventlet==0.39.0
executing==2.1.0
ezdxf==1.4.2
faiss-cpu==1.9.0.post1
fake-useragent==1.5.1
fastapi==0.111.1
fastapi-cli==0.0.7
fastavro==1.10.0
fastjsonschema==2.21.1
filelock==3.16.1
filetype==1.2.0
firebase-admin==6.6.0
FLAML==2.3.3
Flask==2.2.2
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
flatbuffers==24.12.23
folium==0.19.5
fontawesomefree==6.6.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.6
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.77.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.66.0
gotrue==2.11.1
greenlet==3.1.1
groq==0.15.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-health-checking==1.62.3
grpcio-status==1.62.3
grpcio-tools==1.62.3
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
ijson==3.3.0
importlib_metadata==8.4.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
json_repair==0.35.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.4
kubernetes==30.1.0
langchain==0.2.17
langchain-anthropic==0.1.23
langchain-astradb==0.3.5
langchain-aws==0.1.18
langchain-core==0.2.43
langchain-groq==0.1.10
langchain-mistralai==0.1.13
langchain-mongodb==0.1.9
langchain-openai==0.1.25
langchain-pinecone==0.1.2
langchain-text-splitters==0.2.4
langchainhub==0.1.21
langdetect==1.0.9
langfuse==2.57.10
langsmith==0.1.147
litellm==1.51.0
loguru==0.7.3
lxml==5.3.0
magicattr==0.1.6
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.1
matplotlib==3.6.3
matplotlib-inline==0.1.7
mdurl==0.1.2
metal_sdk==2.5.1
metaphor-python==0.1.23
mistune==3.1.0
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
mss==10.0.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.59.8
opencv-python==4.12.0.88
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
optuna==4.1.0
orjson==3.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
passlib==1.7.4
pgvector==0.2.5
pillow==10.4.0
pinecone-client==3.2.2
platformdirs==4.3.6
plyer==2.1.0
portalocker==2.10.1
postgrest==0.19.1
posthog==3.8.3
praw==7.8.1
prawcore==2.4.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.1
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyautogen==0.2.35
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyglet==2.1.6
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.10.1
pyparsing==3.0.9
pypdf==4.3.1
pyperclip==1.9.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-iso639==2024.10.22
python-jose==3.3.0
python-json-logger==3.2.1
python-magic==0.4.27
python-multipart==0.0.7
python-pptx==0.6.23
python-socketio==5.12.1
pytube==15.0.0
pytz==2024.2
pywaffle==1.1.1
pywin32==306
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.13.0
qianfan==0.3.5
RapidFuzz==3.11.0
realtime==2.1.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
rtree==1.4.0
s3transfer==0.10.4
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.20.0
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlmodel==0.0.18
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.37.2
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.2
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triangle==20250106
trimesh==4.6.10
typer==0.12.5
types-cachetools==5.5.0.20240820
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
ultralytics==8.3.167
ultralytics-thop==2.0.14
uncurl==0.0.11
unstructured==0.14.10
unstructured-client==0.29.0
update-checker==0.18.0
upstash-vector==0.4.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uuid6==2024.7.10
uv==0.5.21
uvicorn==0.30.6
validators==0.34.0
virtualenv==20.29.2
watchfiles==1.0.4
wcwidth==0.2.13
weaviate-client==4.9.6
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==2.2.2
wikipedia==1.4.0
win10toast-click==0.1.2
win32_setctime==1.2.0
winotify==1.1.0
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
XlsxWriter==3.2.0
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
youtube-transcript-api==0.6.3
zep-python==2.0.2
zipp==3.21.0
